# Conflicts:
#	FrostyModSupport/Interfaces/IResourceContainer.cs
#	FrostyModSupport/Mod/FrostyMod.cs
#	FrostyModSupport/Mod/FrostyModCollection.cs
#	FrostyModSupport/Mod/FrostyModDetails.cs
#	FrostyModSupport/Mod/Resources/BaseModResource.cs
#	FrostyModSupport/Mod/Resources/BundleModResource.cs
#	FrostyModSupport/Mod/Resources/ChunkModResource.cs
#	FrostyModSupport/Mod/Resources/EbxModResource.cs
#	FrostyModSupport/Mod/Resources/EmbeddedModResource.cs
#	FrostyModSupport/Mod/Resources/FsFileModResource.cs
#	FrostyModSupport/Mod/Resources/ModResourceType.cs
#	FrostyModSupport/Mod/Resources/ResModResource.cs

+++ FrostySdk/IO/Compression/CompressionZStd.cs
@@ -10,23 +10,23 @@ public partial class CompressionZStd : ICompressionFormat
     public string Identifier => "ZStandard";
     private const string NativeLibName = "zstd";
     
-    private static IntPtr s_DDict = IntPtr.Zero;
+    private static nuint s_DDict = nuint.Zero;
     
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_getErrorName(IntPtr code);
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_isError(IntPtr code);
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_createDDict(IntPtr dict, IntPtr dictSize);
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_createDCtx();
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_freeDCtx(IntPtr dctx);
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_decompress(IntPtr dst, IntPtr dstCapacity, IntPtr src, IntPtr compressedSize);
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_decompress_usingDDict(IntPtr dctx, IntPtr dst, IntPtr dstCapacity, IntPtr src, IntPtr srcSize, IntPtr dict);
-    [LibraryImport(NativeLibName)] internal static partial IntPtr ZSTD_compress(IntPtr dst, IntPtr dstCapacity, IntPtr src, IntPtr srcSize);
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_getErrorName(nuint code);
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_isError(nuint code);
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_createDDict(nuint dict, nuint dictSize);
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_createDCtx();
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_freeDCtx(nuint dctx);
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_decompress(nuint dst, nuint dstCapacity, nuint src, nuint compressedSize);
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_decompress_usingDDict(nuint dctx, nuint dst, nuint dstCapacity, nuint src, nuint srcSize, nuint dict);
+    [LibraryImport(NativeLibName)] internal static partial nuint ZSTD_compress(nuint dst, nuint dstCapacity, nuint src, nuint srcSize);
 
     /// <summary>
     /// Checks if the specified code is a valid ZStd error.
     /// </summary>
-    private unsafe static void GetError(IntPtr code)
+    private unsafe static void GetError(nuint code)
     {
-        if (ZSTD_isError(code) != IntPtr.Zero)
+        if (ZSTD_isError(code) != nuint.Zero)
         {
             string error = new((sbyte*)ZSTD_getErrorName(code));
             throw new Exception($"A ZStandard operation failed with error: \"{error}\"");
@@ -36,22 +36,22 @@ public partial class CompressionZStd : ICompressionFormat
     /// <inheritdoc/>
     public unsafe void Decompress<T>(Block<T> inData, ref Block<T> outData, CompressionFlags inFlags = CompressionFlags.None) where T : unmanaged
     {
-        IntPtr code;
+        nuint code;
         if (inFlags.HasFlag(CompressionFlags.ZStdUseDicts))
         {
-            if (s_DDict == IntPtr.Zero)
+            if (s_DDict == nuint.Zero)
             {
                 Block<byte> ebxDict = FileSystemManager.GetFileFromMemoryFs("Dictionaries/ebx.dict");
-                s_DDict = ZSTD_createDDict((IntPtr)ebxDict.Ptr, ebxDict.Size);
+                s_DDict = ZSTD_createDDict((nuint)ebxDict.Ptr, (nuint)ebxDict.Size);
             }
 
-            IntPtr dctx = ZSTD_createDCtx();
-            code = ZSTD_decompress_usingDDict(dctx, (IntPtr)outData.Ptr, outData.Size, (IntPtr)inData.Ptr, inData.Size, s_DDict);
+            nuint dctx = ZSTD_createDCtx();
+            code = ZSTD_decompress_usingDDict(dctx, (nuint)outData.Ptr, (nuint)outData.Size, (nuint)inData.Ptr, (nuint)inData.Size, s_DDict);
             ZSTD_freeDCtx(dctx);
         }
         else
         {
-            code = ZSTD_decompress((IntPtr)outData.Ptr, outData.Size, (IntPtr)inData.Ptr, inData.Size);
+            code = ZSTD_decompress((nuint)outData.Ptr, (nuint)outData.Size, (nuint)inData.Ptr, (nuint)inData.Size);
         }
         GetError(code);
     }
@@ -59,7 +59,7 @@ public partial class CompressionZStd : ICompressionFormat
     /// <inheritdoc/>
     public unsafe void Compress<T>(Block<T> inData, ref Block<T> outData, CompressionFlags inFlags = CompressionFlags.None) where T : unmanaged
     {
-        IntPtr code = ZSTD_compress((IntPtr)outData.Ptr, outData.Size, (IntPtr)inData.Ptr, inData.Size);
+        nuint code = ZSTD_compress((nuint)outData.Ptr, (nuint)outData.Size, (nuint)inData.Ptr, (nuint)inData.Size);
         GetError(code);
     }
 }
